move to maven build
