+++ .gitignore
@@ -1,0 +1,5 @@
+# maven
+/target
+
+# application files
+simulationLog*
